requirements.txt
streamlit
pytesseract
Pillow
transformers
datasets
torch
faiss-cpu
